Update linear.py
When y is equal to b, it's crucial to ensure that we don't mistakenly include (a, b) as one of the lattice points. This is because when y equals b, the point lies on the ellipse, and there should be no tangent at that location.

Nevertheless, this consideration doesn't impact scenarios where b is a floating-point number. It only applies to our specific problem where b equals 2500 times the square root of 5.